+++ app-java/src/main/java/ly/count/java/demo/Sample.java
@@ -132,7 +132,7 @@ public class Sample {
                     //System.out.println("[" + logLevel + "] " + logMessage);
                 }
             })
-            .setEventsBufferSize(1)//this should be only used for testing
+            .setEventQueueSizeToSend(1)//setting queue size to "1" should only be done for testing, unless you have a really good reason to do it
             .setMetricOverride(metricOverride)
             .setApplicationVersion("123.56.h");