+++ sdk-java/src/test/java/ly/count/sdk/java/internal/BackendModeTests.java
@@ -23,7 +23,7 @@ public class BackendModeTests {
         Config cc = new Config("https://try.count.ly", "COUNTLY_APP_KEY");
         //File targetFolder = new File("/Users/zahidzafar/Projects/countly/countly-sdk-java");
 
-        cc.setEventsBufferSize(4).enableBackendMode();
+        cc.setEventQueueSizeToSend(4).enableBackendMode();
 
         File targetFolder = new File("d:\\__COUNTLY\\java_test\\");
         Countly.init(targetFolder, cc);